* Смена алгоритма работы: можно без вреда пропустить ненужные операции   * Модернизация алгоритма "быстрой сборки"   * Добавлена возможность записи в репозиторий     только в ручном режиме после сборки нового релиза (un|stable)   * Памятка при старте

+++ lib.sh
@@ -15,12 +15,13 @@ import() {
 # y 						- согласие на отдельный вопрос
 # Y 						- согласие на все вопросы
 # любой символ 	- отмена операции, прекращение работы
-# @since 0.4.0
+# @since 0.5.0
 ask_YN() {
 	local AMSURE
-  
+
+while true; do
 	if [ -n "$1" ]; then
-		read -n 1 -p "$1
+		read -n 1 -p "${bold}$1${reset}
     $2
     Ваш выбор: " AMSURE
 	else
@@ -28,24 +29,35 @@ ask_YN() {
 	fi
 	echo "" 1>&2
 
+	# Параметр по умолчанию, если есть
+	if [ -n "$3" ]; then
+		AMSURE=${AMSURE:-${3}}
+	fi
+
 	case "$AMSURE" in
-		y|д)
+		y)
 			echo "Сами попросили..."
 			return 0
 			;;
-		Y|Д)
+		Y)
 			echo "Вы согласились на всё сами..."
 			return 1
 			;;
-		*)
-			echo "Отмена..."
-			return 255
+		n)
+			echo "Пропускаем..."
+			return 254
 			;;
-	esac
+		q)
+			echo "Не смогла я... не смогла..."
+			exit 0
+			;;
+		*) echo "Давайте попробуем ещё раз:";;
+		esac
+	done
 }
 
 # Функция проверки статуса выполнения операции
-# @since 0.2.0
+# @since 0.4.0
 check_success() {
 	local STATUS="[ OK ]"
 
@@ -67,7 +79,7 @@ check_success() {
 }
 
 # Собираем пакет
-# @since 0.2.0
+# @since 0.4.0
 build() {
 	dpkg-buildpackage -rfakeroot
 
@@ -76,21 +88,44 @@ build() {
 }
 
 # Ставим пакет
-# @since 0.2.0
+# @since 0.5.0
 install() {
 	if [ -e $1 ]; then
 		sudo dpkg -i $1
 
 		check_success $? "Установка..."
 		return $?
+	else
+		echo -n "${yellow}Отсутствует deb-файл: $1"
+		echo "${reset}"
+		return 1
 	fi
 }
 
 # Удаляем сбилдованный deb-файл
-# @since 0.4.0
+# @since 0.5.0
 deldeb() {
-	rm $*
+	local FILELIST=$(find $pkg_dir -name "${pkginfo}*" -print -quit)
 
-	check_success $? "Удаление файлов сборки..."
-	return $?
+	if [ -n "$FILELIST" ]; then
+		rm $*
+
+		check_success $? "Удаление файлов сборки..."
+		return $?
+	else
+		echo -n "${yellow}Не найдено файлов сборки"
+		echo "${reset}"
+		return 1
+	fi
+}
+
+# Положить пакет в репозиторий
+# @since 0.5.0
+put2rep() {
+	if [ -e $pkg4rep ]; then
+		dput -u $repository_user $pkg4rep
+
+		check_success $? "Запись в репозиторий"
+		return $?
+	fi
 }
